Remove datalocation from the SPIAccessTokenSpec and replace it with SPIAccessTokenDataUpdate "trigger" object that forces the reconcilers to load the data from the storage.
This seems to be the cleanest solution to the problem of synchronizing
the cluster state with out-of-cluster storage without resorting to
periodical polling of all data.

It is intended that the OAuth service/the end user will have the perms
to create the SPIAccessTokenDataUpdate objects and will therefore be able
to use the NotifyingTokenStorage.