feat: Enhance error handling in the operator (#98)
* Improve the error handling in the SPIAccessToken reconciliation.

Introduce Invalid and Error phases, error reason and error message to the
SPIAccessToken status.

Enhance the github metadata provider with the ability to distinguish between
token errors and other "environmental" errors.

* The AuthenticatedOAuthState is not used in operator and is strictly related
to the OAuth service. It is therefore removed from the shared codebase and
the Codec is made more flexible for the callers to be able to (en|de)code
other datastructures.